Test Passes

+++ libclc/generic/include/spirv/image/image.h
@@ -25,7 +25,7 @@ _CLC_OVERLOAD _CLC_DECL float4 __spirv_ImageRead__Rfloat4(
 
 // BINDLESS IMAGES PROTOTYPE
 _CLC_OVERLOAD _CLC_DECL float4 __spirv_ImageRead__Rfloat4(
-    unsigned long imageHandle, float coord);
+    unsigned long imageHandle, int coord);
 
 _CLC_OVERLOAD _CLC_DECL float4 __spirv_ImageSampleExplicitcoord__Rfloat4(
     sampler_t sampler, float coord, int op1, float op2);
